language-model/pyproject.toml
[project]
name = "language-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "google-api-python-client>=2.181.0",
    "google-cloud>=0.34.0",
    "google-cloud-vision>=3.10.2",
    "google-genai>=1.33.0",
    "jinja2>=3.1.6",
    "openai>=1.107.1",
    "pydantic>=2.11.7",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
]
